[javascript/node/manifestSources] remove force-update flag and implementation.
This feature was unused until now and it is not really clear if we'll need
it at all.If we decide later, that we need it, it's probably best implemented
centrally in ManifestServer, rather than in each manifest Source implementation
seperately and maybe incompletely. ManifestServer could genereate a hash over
the files and job message and then decide if it dispatched the job already.
Note also that ManifestMaster is also trying to avoid running duplicate jobs.