[PRDispatcher] WIP browser|socket.io: implement simple reconnect logic.
When the websocket connection between client and server gets
disconnected, the client is unsubscribed by the server and resources are
freed. This happens e.g. with the GKE ingress eventually:

> With HTTP(S) Load Balancing, the WebSocket protocol just works. No
> configuration is required.
>
> If you intend to use the WebSocket protocol, you might want to use a
> timeout value larger than the default 30 seconds. To set the timeout
> value for a backend service configured through Ingress, create a
> BackendConfig object, and use the beta.cloud.google.com/backend-config
> annotation in your Service manifest.

 -- https://cloud.google.com/kubernetes-engine/docs/concepts/ingress#support_for_websocket

also:

> * When sending WebSocket traffic to an HTTP(S) load balancer, the
backend service timeout is interpreted as the maximum amount of time
that a WebSocket, idle or active, can remain open.

 -- https://cloud.google.com/load-balancing/docs/backend-service#backend_service_components

Hence, we can't have Websockets that remain open indefinitely, currently
since the default timeout is 30 secconds, we can see disconnections very
frequently.

The clients can, as it seems, all handle a fresh subscription upon old
state, hence these simple fixes "seem" to be sufficient so far.